+++ app/urls.py
@@ -15,6 +15,7 @@ urlpatterns = [
     path("tags/", include("www.tags.urls")),
     path("contributions/", include("www.contributions.urls")),
     path("glossary/", include("www.glossary.urls")),
+    path("admin/", include("www.admin.urls")),
     # admin
     path("django/", admin_site.urls),
     # api
