+++ templates/profile/home.html
@@ -55,29 +55,5 @@
             </div>
         </div>
     </div>
-
-    {% if user.has_role_administrator %}
-        <br />
-        <br />
-        <h4>Administration</h4>
-        <div class="row mb-3">
-            <div class="col-md-4">
-                <div class="card">
-                    <div class="card-body">
-                        <h5 class="card-title">Contributeurs <small class="text-muted">{{ contributor_count }}</small></h5>
-                        <a class="btn btn-primary" href="{% url 'profile:admin_contributors' %}">Voir</a>
-                    </div>
-                </div>
-            </div>
-            <div class="col-md-4">
-                <div class="card">
-                    <div class="card-body">
-                        <h5 class="card-title">Historique</h5>
-                        <a class="btn btn-primary" href="{% url 'profile:admin_history' %}">Voir</a>
-                    </div>
-                </div>
-            </div>
-        </div>
-    {% endif %}
 </div>
 {% endblock %}
